Keep track of when very first buffer is played
Pretty useless right now as StreamInstant is basically private all over.
The only thing it can realistically do is determine a Duration based on
another StreamInstant, which you can't make yourself, but I guess it
might come in handy? Either way, it's been a cause of frustration for a
while.